Migrate licence data files to single-object shape
Starts moving data files off the legacy array-of-one convention, which forced every consumer to destructure a single item out of an array purely to reach it. licence.data.js is split into one file per entity (licence, permit licence, licence document, licence document header, licence document role, licence version), each returning the bare entity with no table-name wrapper; a new _licence() helper in licence.scenario.js composes them. The load fixture now pluralizes singular keys and array-wraps values in one place, so data, scenario, and spec files can compose and destructure using plain singular entity names throughout. Updated the six spec files that build on licence.scenario.js to match.
