+++ tests/internal/charge-information/create-replace-then-add-charge-versions.spec.js
@@ -8,11 +8,7 @@ test.describe('Create, replace then add a charge version (internal)', { tag: '@s
   test.beforeAll(async ({ setup }) => {
     const scenario = scenarioData()
 
-    const {
-      licences: [scenarioLicence]
-    } = scenario
-
-    licence = scenarioLicence
+    licence = scenario.licence
 
     await setup(scenario)
   })
