+++ tests/internal/return-versions/submit-no-returns-required-journey.spec.js
@@ -8,11 +8,7 @@ test.describe('Submit no returns requirement (internal)', () => {
   test.beforeAll(async ({ setup }) => {
     const scenario = scenarioData()
 
-    const {
-      licences: [scenarioLicence]
-    } = scenario
-
-    licence = scenarioLicence
+    licence = scenario.licence
 
     await setup(scenario)
   })
