+++ tests/support/data/permit-licence.data.js
@@ -1,0 +1,9 @@
+export default function (licence) {
+  return {
+    licenceRef: licence.licenceRef,
+    startDate: licence.startDate,
+    metadata: {
+      source: 'acceptance-test-setup'
+    }
+  }
+}
